requirements.txt
argon2-cffi==16.2.0
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.16.1
awsebcli==3.9.1
botocore==1.5.9
cement==2.8.2
cffi==1.8.3
channels==0.17.3
colorama==0.3.7
constantly==15.1.0
daphne==0.15.0
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.10.1
django-allauth==0.27.0
django-appconf==1.0.2
django-autocomplete-light==3.2.1
django-compressor==2.1
django-crispy-forms==1.6.0
django-file-form==0.2.0
django-filter==0.15.0
django-sass-processor==0.4.6
djangorestframework==3.4.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
incremental==16.10.1
jmespath==0.9.1
libsass==0.11.1
Markdown==2.6.6
msgpack-python==0.4.8
oauthlib==2.0.0
pathlib==1.0.1
pathspec==0.5.0
Pillow==3.3.1
psycopg2==2.6.2
pycparser==2.14
Pygments==2.1.3
python-dateutil==2.6.0
python3-openid==3.0.10
pytz==2016.7
PyYAML==3.12
rcssmin==1.0.6
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.2
rjsmin==1.0.12
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
Twisted==16.5.0
txaio==2.5.2
websocket-client==0.40.0
zope.interface==4.3.2